refactor(cordova): use build runner for build/prepare
